+++ Mods ChoGGi/Disable Building Maintenance Cycle/metadata.lua
@@ -12,9 +12,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "3042801784",
 	"pops_any_uuid", "2ee0ad6b-b88a-4c89-9101-e9471f4d9362",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {